+++ src/player/main/sp/it/pl/ui/objects/picker/ContainerPicker.kt
@@ -6,6 +6,7 @@ import javafx.geometry.Orientation.VERTICAL
 import sp.it.pl.layout.Container
 import sp.it.pl.layout.ContainerBi
 import sp.it.pl.layout.ContainerFreeForm
+import sp.it.pl.layout.ContainerSeq
 import sp.it.pl.main.APP
 import sp.it.pl.main.IconUN
 import sp.it.pl.main.emScaled
@@ -23,6 +24,8 @@ class ContainerPicker(onContainerSelect: (Container<*>) -> Unit, onWidgetSelect:
          sequenceOf(
             CellData("Split Vertically", IconUN(0x2ff0), "Splits space to left and right layout.") { onContainerSelect(ContainerBi(HORIZONTAL)) },
             CellData("Split Horizontally", IconUN(0x2ff1), "Splits space to top and bottom layout.") { onContainerSelect(ContainerBi(VERTICAL)) },
+            CellData("Sequence Horizontal", IconUN(0x2ff2), "Splits space to top and bottom layout.") { onContainerSelect(ContainerSeq(VERTICAL)) },
+            CellData("Sequence Vertical", IconUN(0x2ff3), "Splits space to top and bottom layout.") { onContainerSelect(ContainerSeq(VERTICAL)) },
             CellData("FreeForm", IconUN(0x2ffb), "Free form layout. Components behave like windows.") { onContainerSelect(ContainerFreeForm()) }
          ) + (
             if (APP.widgetManager.widgets.separateWidgets.value) {